Move basic views to new module
