+++ app/build.gradle
@@ -165,6 +165,7 @@ dependencies {
         System.out.println("app: free build hack, skipping some dependencies")
     }
     implementation project(':ui:app-start-intent')
+    implementation project(':ui:common')
 
     annotationProcessor "androidx.annotation:annotation:$annotationVersion"
     implementation "androidx.appcompat:appcompat:$appcompatVersion"
