+++ app/src/main/res/layout/filter_dialog_row.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<de.danoeh.antennapod.view.RecursiveRadioGroup
+<de.danoeh.antennapod.ui.common.RecursiveRadioGroup
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
@@ -58,4 +58,4 @@
         android:layout_gravity="center_vertical"
         android:checked="true" />
 
-</de.danoeh.antennapod.view.RecursiveRadioGroup>
+</de.danoeh.antennapod.ui.common.RecursiveRadioGroup>
