+++ app/src/main/res/layout/quick_feed_discovery.xml
@@ -34,7 +34,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content">
 
-        <de.danoeh.antennapod.view.WrappingGridView
+        <de.danoeh.antennapod.ui.common.WrappingGridView
                 android:id="@+id/discover_grid"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
