+++ app/src/main/res/layout/quick_feed_discovery_item.xml
@@ -7,7 +7,7 @@
         android:padding="4dp"
         android:clipToPadding="false">
 
-    <de.danoeh.antennapod.view.SquareImageView
+    <de.danoeh.antennapod.ui.common.SquareImageView
             android:id="@+id/discovery_cover"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
