+++ app/src/main/res/layout/secondary_action.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
              xmlns:tools="http://schemas.android.com/tools"
+             xmlns:app="http://schemas.android.com/apk/res-auto"
              android:layout_width="48dp"
              android:layout_height="48dp"
              android:layout_marginRight="12dp"
@@ -19,9 +20,10 @@
             tools:ignore="ContentDescription"
             tools:src="@sample/secondaryaction"/>
 
-    <de.danoeh.antennapod.view.CircularProgressBar
+    <de.danoeh.antennapod.ui.common.CircularProgressBar
             android:id="@+id/secondaryActionProgress"
             android:layout_width="40dp"
             android:layout_gravity="center"
-            android:layout_height="40dp"/>
+            android:layout_height="40dp"
+            app:foregroundColor="?attr/action_icon_color"/>
 </FrameLayout>
