+++ app/src/main/res/layout/subscription_item.xml
@@ -8,7 +8,7 @@
     android:layout_height="match_parent"
     android:foreground="?attr/selectableItemBackground">
 
-    <de.danoeh.antennapod.view.SquareImageView
+    <de.danoeh.antennapod.ui.common.SquareImageView
         android:id="@+id/imgvCover"
         android:layout_width="fill_parent"
         android:layout_height="fill_parent"
