+++ settings.gradle
@@ -1,3 +1,4 @@
 include ':app'
 include ':core'
 include ':ui:app-start-intent'
+include ':ui:common'
