+++ ui/common/README.md
@@ -1,0 +1,3 @@
+# :ui:common
+
+This module provides basic UI functionality that is needed in multiple modules. UI elements that are only used in a single module should not be defined here.
