+++ ui/common/src/main/java/de/danoeh/antennapod/ui/common/WrappingGridView.java
@@ -1,4 +1,4 @@
-package de.danoeh.antennapod.view;
+package de.danoeh.antennapod.ui.common;
 
 import android.content.Context;
 import android.util.AttributeSet;
