+++ ui/common/src/main/res/values/styleable.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <declare-styleable name="SquareImageView">
+        <attr name="direction" format="enum">
+            <enum name="width" value="0"/>
+            <enum name="height" value="1"/>
+            <enum name="minimum" value="2"/>
+        </attr>
+    </declare-styleable>
+
+    <declare-styleable name="CircularProgressBar">
+        <attr name="foregroundColor" format="color" />
+    </declare-styleable>
+
+    <declare-styleable name="PlaybackSpeedIndicatorView">
+        <attr name="foregroundColor" /> <!-- format omitted to avoid double definition -->
+    </declare-styleable>
+</resources>
